Data Redistribution Implementation (Partial)

+++ source/code/inc/pmController.h
@@ -43,7 +43,7 @@ class pmController : public pmBase
 		pmStatus WaitForTaskCompletion_Public(pmTaskHandle pTaskHandle);
 		pmStatus GetTaskExecutionTimeInSecs_Public(pmTaskHandle pTaskHandle, double* pTime);
 		pmStatus SubscribeToMemory_Public(pmTaskHandle pTaskHandle, ulong pSubtaskId, bool pIsInputMemory, pmSubscriptionInfo pSubscriptionInfo);
-        pmStatus RedistributeData_Public(pmTaskHandle pTaskHandle, unsigned long pSubtaskId, size_t pOffset, size_t pLength, unsigned long pOrder);
+        pmStatus RedistributeData_Public(pmTaskHandle pTaskHandle, unsigned long pSubtaskId, size_t pOffset, size_t pLength, unsigned int pOrder);
 		pmStatus SetCudaLaunchConf_Public(pmTaskHandle pTaskHandle, unsigned long pSubtaskId, pmCudaLaunchConf& pCudaLaunchConf);
 
 		uint GetHostId_Public();
